Fix vendor ID type mismatch
Resumen

El repositorio de proveedores ahora maneja los ID como largos, lo que garantiza búsquedas consistentes con la entidad VendorDto

VendorService recupera tiendas utilizando parámetros Long, haciendo coincidir el repositorio y eliminando las discrepancias de tipos

Los puntos finales de VendorController aceptan variables de ruta larga para los ID de proveedores, lo que permite la correcta persistencia de las credenciales de Shopify

Los controladores y servicios convierten los ID de enteros de usuarios a largos antes de consultar a los proveedores, lo que evita errores de conversión.

+++ backend/src/main/java/com/rocket/service/service/VendorService.java
@@ -39,9 +39,9 @@ public class VendorService {
 		return vendorRepository.save(tienda);
 	}
 	
-	public VendorDto obtenerTiendaPorId(Integer id) {
+        public VendorDto obtenerTiendaPorId(Long id) {
 
-		List<VendorDto> tiendas = mongoOperations.find(Query.query(Criteria.where("_id").is(id)), VendorDto.class);
+                List<VendorDto> tiendas = mongoOperations.find(Query.query(Criteria.where("_id").is(id)), VendorDto.class);
 
 		if (tiendas.isEmpty()) {
 			return new VendorDto();
